requirements.txt
torch==2.2.1
transformers==4.39.3
peft==0.10.0
pandas==2.2.2
scikit-learn==1.4.2
spacy==3.7.4
contractions==0.1.73
PyYAML==6.0.1
tqdm==4.66.2